Changed organized_raw to not create links for nonexistent runs; changed qsub scripts for functional runs

+++ organize_raw.sh
@@ -19,8 +19,8 @@ for subj in $(ls -d *); do
 		for r in $(seq 1 24); do
 			i=$(($r-1))
 	
-			if [ ! -L  ${WD}/${subj}/run${r} ]; then
-			
+			if [ $i -le ${#BOLD_DIRS[@]} ] && [ ! -L ${WD}/${subj}/run${r} ]; then
+				echo ${BOLD_DIRS[i]}
 				ln -s ${BOLD_DIRs[i]} run${r}
 				
 			fi
